refactor: extract circle constructions to circle/CircleConstructions.ts
Moves 6 circle construction classes to circle/CircleConstructions.ts.
Constructions.ts: 1057 → 969 lines.

Note: CircumcircleConstruction stays in point/PointConstructions.ts
because CircumcenterConstruction extends it directly.
