+++ src/elements/Constructions.ts
@@ -313,106 +313,18 @@ import {LineConnectConstruction,
     } from "./line/LineConstructions";
 
 
+
 /************************
  * Element Class Circle *
+ * (see circle/CircleConstructions.ts for circle construction classes)
+ * (CircumcircleConstruction lives in point/PointConstructions.ts)
  ************************/
 
-// circle
-// radius
-// points A, B [plane C=screen]
-// the circle with center A and radius AB in the plane C
-// (Java: Slate.java circle case 0, choice 0 — new CircleElement(A,B,screen))
-export class CircleRadiusCenterConstruction extends Construction {
-    constructionMethod: AllConstructions = CircleConstructions.radius;
-    signature: ConstructionTypes[] = [ct.PointElement, ct.PointElement];
+import {CircleRadiusCenterConstruction, CircleRadius3dConstruction,
+    CircleRadius3Point3dConstruction, CircleRadius3PointConstruction,
+    InvertCircleConstruction, SphereIntersectionConstruction
+    } from "./circle/CircleConstructions";
 
-    construct(screen : PlaneElement, params: any[]): [GeomElementsForUpdate, GeomElement] {
-        let ps : PointElement[] = params;
-        let g = new CircleElement({C:ps[0], B:ps[1], AP:screen});
-        return [[g], g];
-    }
-}
-
-// circle — radius (3D, 2-point)
-// points A, B, plane C
-// the circle with center A and radius AB in the plane C
-// (Java: Slate.java circle case 0, choice 2 — new CircleElement(A,B,C))
-export class CircleRadius3dConstruction extends Construction {
-    constructionMethod: AllConstructions = CircleConstructions.radius;
-    signature: ConstructionTypes[] = [ct.PointElement, ct.PointElement, ct.PlaneElement];
-
-    construct(_screen: PlaneElement, params: any[]): [GeomElementsForUpdate, GeomElement] {
-        let ps : PointElement[] = params;
-        let g = new CircleElement({C: ps[0], B: ps[1], AP: params[2] as PlaneElement});
-        return [[g], g];
-    }
-}
-
-// circle — radius (3D, 3-point)
-// points A, B, C, plane D
-// (Java: Slate.java circle case 0, choice 3 — new CircleElement(A,B,C,D))
-export class CircleRadius3Point3dConstruction extends Construction {
-    constructionMethod: AllConstructions = CircleConstructions.radius;
-    signature: ConstructionTypes[] = [ct.PointElement, ct.PointElement, ct.PointElement, ct.PlaneElement];
-
-    construct(_screen: PlaneElement, params: any[]): [GeomElementsForUpdate, GeomElement] {
-        let ps : PointElement[] = params;
-        let g = new CircleElement({C: ps[0], A: ps[1], B: ps[2], AP: params[3] as PlaneElement});
-        return [[g], g];
-    }
-}
-
-// circle — radius (2D, 3-point)
-// points A, B, C
-// the circle with center A and radius |BC| in the screen plane
-// (Java: CircleElement(A, B, C, screen) — A=center, radius=B.distance(C))
-// MUST be registered BEFORE the 2-point variant in the constructions array
-// (signature variant ordering rule: longer signature first)
-export class CircleRadius3PointConstruction extends Construction {
-    constructionMethod: AllConstructions = CircleConstructions.radius;
-    signature: ConstructionTypes[] = [ct.PointElement, ct.PointElement, ct.PointElement];
-
-    construct(screen : PlaneElement, params: any[]): [GeomElementsForUpdate, GeomElement] {
-        let ps : PointElement[] = params;
-        let g = new CircleElement({C: ps[0], A: ps[1], B: ps[2], AP: screen});
-        return [[g], g];
-    }
-}
-
-
-// circle
-// invert
-// circles A, B
-// the image of circle A inverted in circle B
-// (Java: InvertCircle.java — TS: InvertCircleElement.ts)
-export class InvertCircleConstruction extends Construction {
-    constructionMethod: AllConstructions = CircleConstructions.invert;
-    signature: ConstructionTypes[] = [ct.CircleElement, ct.CircleElement];
-
-    construct(screen: PlaneElement, params: any[]): [GeomElementsForUpdate, GeomElement] {
-        const C = params[0] as CircleElement;
-        const D = params[1] as CircleElement;
-        const g = new InvertCircleElement(C, D);
-        return [[g], g];
-    }
-}
-
-// circle
-// intersection
-// spheres A, B
-// the circle at the intersection of spheres A and B
-// (Java: IntersectionSS.java — TS: SphereIntersectionElement.ts)
-export class SphereIntersectionConstruction extends Construction {
-    constructionMethod: AllConstructions = CircleConstructions.intersection;
-    signature: ConstructionTypes[] = [ct.SphereElement, ct.SphereElement];
-
-    construct(screen: PlaneElement, params: any[]): [GeomElementsForUpdate, GeomElement] {
-        const S = params[0] as SphereElement;
-        const T = params[1] as SphereElement;
-        const g = new SphereIntersectionElement(S, T);
-        return [[g], g];
-    }
-}
 
 /*************************
  * Element Class Polygon *
